FIM nodes_count database variable is not checking correctly

+++ src/syscheckd/db/fim_db_registries.c
@@ -458,7 +458,7 @@ int fim_db_insert_registry_data(fdb_t *fim_sql,
             return FIMDB_ERR;
         }
 
-        if ((unsigned int)count >= syscheck.file_limit) {
+        if (count >= syscheck.file_limit) {
             mdebug1("Couldn't insert '%s' value entry into DB. The DB is full, please check your configuration.",
                     data->name);
             return FIMDB_FULL;
@@ -488,7 +488,7 @@ int fim_db_insert_registry_key(fdb_t *fim_sql, fim_registry_key *entry, unsigned
             return FIMDB_ERR;
         }
 
-        if ((unsigned int)count >= syscheck.file_limit) {
+        if (count >= syscheck.file_limit) {
             mdebug1("Couldn't insert '%s %s' entry into DB. The DB is full, please check your configuration.",
                     registry_arch[entry->arch], entry->path);
             return FIMDB_FULL;
